Fix circular import issue and created departmentPositions file.
