+++ app/logic/tracy.py
@@ -4,7 +4,12 @@ from app.models.Tracy import db
 from app.models.Tracy.stuposn import STUPOSN
 from app.models.Tracy.studata import STUDATA
 from app.models.Tracy.stustaff import STUSTAFF
+from app.models.student import Student
 from app import app
+
+
+class InvalidUserException(Exception):
+    pass
 
 class InvalidQueryException(Exception):
     pass
@@ -148,3 +153,48 @@ class Tracy():
             student = self.getStudentFromEmail(email)
             if student:
                 return "Student"
+
+
+def createStudentFromTracy(username=None, bnumber=None):
+    """
+        Attempts to add a student from the Tracy database to the application, based on the provided username or bnumber.
+
+        Raises InvalidUserException if this does not succeed.
+    """
+    if not username and not bnumber:
+        raise ValueError("No arguments provided to createStudentFromTracy()")
+
+    if bnumber:
+        try:
+            tracyStudent = Tracy().getStudentFromBNumber(bnumber)
+        except InvalidQueryException as e:
+            raise InvalidUserException("{} not found in Tracy database".format(bnumber))
+
+    else:    # Executes if no ID is provided
+        email = "{}@berea.edu".format(username)
+        try:
+            tracyStudent = Tracy().getStudentFromEmail(email)
+        except InvalidQueryException as e:
+            raise InvalidUserException("{} not found in Tracy database".format(email))
+
+    # Create the student in Tracy
+    try:
+        return Student.get(Student.ID == tracyStudent.ID.strip())
+    except DoesNotExist:
+        #print('Could not find {0} {1} in Student table, creating new entry.'.format(tracyStudent.FIRST_NAME, tracyStudent.LAST_NAME))
+        return Student.create(ID = tracyStudent.ID.strip(),
+                            PIDM = tracyStudent.PIDM,
+                            legal_name = tracyStudent.FIRST_NAME,
+                            LAST_NAME = tracyStudent.LAST_NAME,
+                            CLASS_LEVEL = tracyStudent.CLASS_LEVEL,
+                            ACADEMIC_FOCUS = tracyStudent.ACADEMIC_FOCUS,
+                            MAJOR = tracyStudent.MAJOR,
+                            PROBATION = tracyStudent.PROBATION,
+                            ADVISOR = tracyStudent.ADVISOR,
+                            STU_EMAIL = tracyStudent.STU_EMAIL,
+                            STU_CPO = tracyStudent.STU_CPO,
+                            LAST_POSN = tracyStudent.LAST_POSN,
+                            LAST_SUP_PIDM = tracyStudent.LAST_SUP_PIDM)
+    else:
+        raise InvalidUserException("Error: Could not get or create {0} {1}".format(tracyStudent.FIRST_NAME, tracyStudent.LAST_NAME))
+
